Move functions methods to the intp::functions namespace
