+++ libsysinspect/src/intp/actions.rs
@@ -1,12 +1,10 @@
 use super::{
     actproc::{modfinder::ModCall, response::ActionResponse},
     constraints::Expression,
+    functions,
     inspector::SysInspector,
 };
-use crate::{
-    util::{dataconv, func},
-    SysinspectError,
-};
+use crate::{util::dataconv, SysinspectError};
 use colored::Colorize;
 use serde::{Deserialize, Serialize};
 use serde_yaml::Value;
@@ -105,7 +103,8 @@ impl Action {
     ) -> Result<Vec<Expression>, SysinspectError> {
         let mut out: Vec<Expression> = Vec::default();
         for mut expr in v_expr {
-            if let Some(modfunc) = func::is_function(&dataconv::to_string(expr.get_op()).unwrap_or_default()).ok().flatten() {
+            if let Some(modfunc) = functions::is_function(&dataconv::to_string(expr.get_op()).unwrap_or_default()).ok().flatten()
+            {
                 match inspector.call_function(eid, &state, &modfunc) {
                     Ok(Some(v)) => expr.set_active_op(v)?,
                     Ok(_) => {}
@@ -174,7 +173,7 @@ impl Action {
 
             for (kw, arg) in &mod_args.args() {
                 let mut arg = arg.to_owned();
-                if let Ok(Some(func)) = func::is_function(&arg) {
+                if let Ok(Some(func)) = functions::is_function(&arg) {
                     match inspector.call_function(eid, &modcall.state(), &func) {
                         Ok(None) => {
                             return Err(SysinspectError::ModelDSLError(format!(
