+++ libsysinspect/src/intp/actproc/modfinder.rs
@@ -3,8 +3,9 @@ use crate::{
     intp::{
         actproc::response::ConstraintFailure,
         constraints::{Constraint, ConstraintKind},
+        functions,
     },
-    util::{dataconv, func},
+    util::dataconv,
     SysinspectError,
 };
 use core::str;
@@ -117,7 +118,7 @@ impl ModCall {
         }
 
         for exp in exp {
-            let fact = func::get_by_namespace(resp.data(), &exp.get_fact_namespace());
+            let fact = functions::get_by_namespace(resp.data(), &exp.get_fact_namespace());
             let res = exp.eval(fact.to_owned());
             if !res.is_positive() {
                 let mut traces: Vec<String> =
@@ -139,7 +140,7 @@ impl ModCall {
 
         let mut traces: Vec<String> = vec![];
         for exp in exp {
-            let res = exp.eval(func::get_by_namespace(resp.data(), &exp.get_fact_namespace()));
+            let res = exp.eval(functions::get_by_namespace(resp.data(), &exp.get_fact_namespace()));
             if res.is_positive() {
                 return (Some(true), None);
             }
@@ -157,7 +158,7 @@ impl ModCall {
         }
 
         for e in exp {
-            let fact = func::get_by_namespace(resp.data(), &e.get_fact_namespace());
+            let fact = functions::get_by_namespace(resp.data(), &e.get_fact_namespace());
             let res = e.eval(fact.to_owned());
             if res.is_positive() {
                 let mut traces: Vec<String> =
