+++ libsysinspect/src/util/mod.rs
@@ -1,2 +1,1 @@
 pub mod dataconv;
-pub mod func;
